refactor(scripting): extract Lua system/job bridge logic to system_bridge.rs
