+++ engine/core/src/scripting/mod.rs
@@ -23,4 +23,5 @@ pub use engine::ScriptEngine;
 
 pub mod event_bus;
 pub mod helpers;
+pub mod system_bridge;
 pub mod worldgen_bridge;
